# Optimize query string parsing