#### QUALITY CHECKLIST
- [X] Changes implemented in code
- [x] Complies with our [contributing guidelines](https://github.com/json-api-dotnet/JsonApiDotNetCore/blob/master/.github/CONTRIBUTING.md)
- [x] Adapted tests
- [ ] Documentation updated

While profiling the query string parsing, 40% of the time was spent in the GC 😬 so I focused on optimizing allocations.

Before
| Method  | argument            | Mean      | Error     | StdDev    | Gen0   | Gen1   | Allocated |
|-------- |-------------------- |----------:|----------:|----------:|-------:|-------:|----------:|
| ReadAll | filt(...)name [121] | 12.347 us | 0.0965 us | 0.2099 us | 0.0610 | 0.0153 |  19.92 KB |

After
| Method  | argument            | Mean      | Error     | StdDev    | Gen0   | Gen1   | Allocated |
|-------- |-------------------- |----------:|----------:|----------:|-------:|-------:|----------:|
| ReadAll | filt(...)name [121] | 10.014 us | 0.1852 us | 0.4143 us | 0.0458 | 0.0153 |  15.84 KB |

You can find the result of each commit in their message.