@bkoelman 2024-02-28
I don't mean to discourage you, but these micro-optimizations do not make much sense to me. Aside from the logging (2ms), we're talking about a max gain of 1ms, which is neglectable in a typical API request involving a network call to a database server. We're not CoreCLR.

Also, the benchmarks in the codebase are "whatever somebody was once curious about", like a playground. It's not a suite of scenarios that represent typical usage. The PerformanceReports repo is better aligned with that.

For logging, its source generator should be used instead, which avoid formatting the message. It's on my list of potential performance improvements. Though it's low priority, compared to other candidates with likely much higher gain. One of the biggest is to look into `dynamic` and reflection usage, replacing them with cached precompiled delegates.

For `Get[Attributes/Relationships]InTypeOrDerived`, it could be evaluated lazily instead of using a concurrent dictionary. That optimizes throughput at the cost of higher startup time, which is a good trade-off.

I'm not against replacing `.First()` with `[0]` when we know there's at least one element, or `.Any()` with `.Count == 0`, but the codebase of full of them, so fixing just one isn't valuable. Again, this is an improvement barely measurable. I'd rather focus on more important things.

The same for `dictionary.TryGetValue`: unless it's reasonable to assume there may be thousands of entries, it doesn't pay off. Inside Dictionary, there's a threshold where it switches from an array to a tree. As array lookups are vectorized these days, the change may not even result in a measurable gain.