@verdie-g 2024-02-28
> I don't mean to discourage you, but these micro-optimizations do not make much sense to me. Aside from the logging (2ms), we're talking about a max gain of 1ms, which is neglectable in a typical API request involving a network call to a database server. We're not CoreCLR.

The optimizations made don't target the execution time but the pressure the code puts on the GC. That JADNC middleware execution time will always be negligible compared to the SQL query. However, by allocating too much you can very easily impact the tail latencies of an application and allocating 20KB to parse 100 bytes string is **huge**. This is what I'm targeting here, and it easily reduces the allocations by 20% of the query parsing.

> For Get[Attributes/Relationships]InTypeOrDerived, it could be evaluated lazily instead of using a concurrent dictionary. That optimizes throughput at the cost of a higher latency for the first request, which is a good trade-off.

I actually wanted to eagerly evaluate them but doing it lazily is fine too. The proposed implementation using the ConcurrentDictionary is doing that.

> I'm not against replacing .First() with [0] when we know there's at least one element, or .Any() with .Count == 0, but the codebase of full of them, so fixing just one isn't valuable. Again, this is an improvement barely measurable. I'd rather focus on more important things.

For those it just hurts me to see an enumerator allocated just to get a length and they are so easy to fix.

> The same for dictionary.TryGetValue

Same idea, it was in the way, it doesn't save an allocation but it was easy to fix.

> Inside Dictionary, there's a threshold where it switches from an array to a tree

I double-checked but I don't think there is such complex algorithm in [Dictionary](https://github.com/dotnet/runtime/blob/main/src/libraries/System.Private.CoreLib/src/System/Collections/Generic/Dictionary.cs). It's a simple hash table algorithm.