@verdie-g 2024-02-28
Though I agree I'm definitely not taking the problem by the right end. The steps were just:
1. Oh there is actually a benchmark project?
2. 20KB allocated that sounds off
3. Easy quick wins

I haven't got the time to look at #1472 but I will some day.