@bkoelman 2024-02-28
Ok, so the 20 KB allocation is reduced down to 16 KB, saving 4 KB. As I said, there are many things more important to worry about. JADNC is stateless, it's easy to scale up extra instances. We never had performance complaints, so working on this is not a priority. It requires effort from both of us that is better spent on things that users _do_ care about.

ConcurrentDictionary lookups are slower than lookups from a regular dictionary, so what you did is not the same. But I don't want to even discuss that.

I'm not going to review this PR, so I suggest you close it and not waste time on it.